fix: preserve fractional zone minutes in custom targets
- Instead of converting accumulated zone seconds into integer minutes (which truncates any leftover seconds), it now preserves them as `Double` values.
- Matches `HistoryAnalyzer`'s `getCustomTargetValue` which already deserializes JSON values as `Double`.

+++ app/src/main/java/com/chrislentner/coach/health/HealthConnectSyncRoutine.kt
@@ -124,9 +124,9 @@ class HealthConnectSyncRoutine(
                 }
             }
 
-            val customTargetsMap = mutableMapOf<String, Int>()
+            val customTargetsMap = mutableMapOf<String, Double>()
             for ((tgt, secs) in customTargetsMapSecs) {
-                val mins = (secs / 60).toInt()
+                val mins = secs / 60.0
                 if (mins > 0) {
                     customTargetsMap[tgt] = mins
                 }
